backend/requirements.txt
google-genai
fastapi
pydantic
uvicorn
python-dotenv
uuid
